[BROKEN] Start unification, problematic
Previously I chose the structure: 'a net SItab for the rest of the tree.
That is, to get to the next net the string and int key are required
simultaneously. This is disadvantageous because for unification we must
match arbitrary terms (strings) to variables. A type of 'a net Inttab
Symtab would allow one to to Symtab.dest |> foldl (handle_arg) [] or
something like that. That change is fairly invasive and prone to errors,
therefore this commit.