Modify random number generator
Choose different a according to wikipedia/authors recommendation
Use int instead of float. In theory the float has a high enough
precision (53 bit) but there is no reason to use floats as we only use
div/mod on the seeds. According to
https://smlfamily.github.io/Basis/integer.html#INTEGER:SIG:SPEC
Ints check for overflows so that should be an additional safeguard.
Superficial testing indicates better random distribution but that was
just a quick test.