Huge commit. Details inside
Rename Path to PathIndex and Net to NetIndex to avoid name clashes
Rework indexgen / netgen
Use Time.time instead of Timing.timing. Basically: only use cpu timing
instead of the three separate ones
Rework term gen to not place a var at the root
Rework benchmarks to share the seeds instead of generating all indices
at the start. Saves a lot of RAM, but is slower.
Use a list of (size,runs) for benchmarks.
Fix printing issues. Every component of a benchmark introduces one
tag (e.g. Size, Generator, Test). All results are stored in one large
list. Queries are done by specifying the x- and y-labels. This works
rather simply: Specifying (Gen "V") labels this axis with all Generators
that are a superstring of V. Also works with an empty string.
As there are multiple values remaining for each cell of the table we
have an additional printer parameter that turns this list of tags +
result into a final string. With the provided helper function (compare)
we simply specify only: Name, x-label, y-label and a list of tags. This
list is used to retain only the tags that are a superset of these tags.
These are then summed. If Size is not given it is implicitly assumed to
be Size "". Therefore, by default every run is summed across different
sizes. An error is raised if some label has not been used for x,y, or selection
By specifying e.g. (Index "") (Test "") [Gen "LV", Size "50"] we get a
table with all indices on the x axis, all tests as y axis / lines and
the only the respective run with the LV generator and size 50 as
entries.
In retrospect: This benchmarking "suite" was a mistake: Extremely time
consuming to implement, somewhat complex, a little bit awkward to use
and a little prone to errors (e.g. accidental sharing of substrings of
tags, although that is basically the only issue as far as I know). Essentially this is a self-built database together with queries.
Now only to introduce the joins of multiple tables, a more fully
featured query language, perhaps Standard ML Query Language, SQL? ...
As of now Path Indexing is still !multiple times! faster than the
variant with Termtables for inexplicable reasons. I have yet to find the reason